+++ lib/avromatic/rspec.rb
@@ -1,5 +1,5 @@
 require 'webmock/rspec'
-require 'avro_turf/test/fake_schema_registry_server'
+require 'avromatic/test/fake_schema_registry_server'
 
 RSpec.configure do |config|
   config.before(:each) do